requirements.txt
ago==0.0.92
alabaster==0.7.12
alembic==1.0.0
amqp==2.3.2
arrow==0.12.1
Babel==2.6.0
billiard==3.5.0.4
binaryornot==0.4.4
celery==4.2.1
certifi==2018.8.24
Chameleon==3.4
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
docutils==0.14
-e git+https://github.com/qlands/elasticfeeds.git@f128339c02f73285fb0d96c636d3975d3b3f86d8#egg=elasticfeeds
elasticsearch==6.3.1
FormEncode==1.3.1
future==0.16.0
gevent==1.3.6
greenlet==0.4.15
gunicorn==19.9.0
hupper==1.3
idna==2.7
imagesize==1.1.0
inflect==1.0.0
Jinja2==2.10
jinja2-time==0.2.0
kombu==4.2.1
linecache2==1.0.0
lingua==4.13
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
mysql-connector-python==8.0.12                                                                                                                                                            
nose==1.3.7                                                                                                                                                                               
packaging==18.0                                                                                                                                                                           
PasteDeploy==1.5.2                                                                                                                                                                        
pkg-resources==0.0.0                                                                                                                                                                      
plaster==1.0                                                                                                                                                                              
plaster-pastedeploy==0.6                                                                                                                                                                  
polib==1.1.0                                                                                                                                                                              
poyo==0.4.1                                                                                                                                                                               
protobuf==3.6.1                                                                                                                                                                           
pycrypto==2.6.1                                                                                                                                                                           
Pygments==2.2.0
pyparsing==2.3.0
pyramid==1.9.2
pyramid-debugtoolbar==4.5
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pyramid-retry==0.5
pyramid-tm==2.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
PyUtilib==5.4.1
-e git+https://github.com/XLSForm/pyxform.git@f00ec761c989c7f074d757d0b7f03577dda2927d#egg=pyxform
redis==2.10.6
repoze.lru==0.7
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.12
stream-framework==1.4.0
test1==1.0
traceback2==1.4.0
transaction==2.2.1
translationstring==1.3
unicodecsv==0.14.1
unittest2==1.1.0
urllib3==1.23
validate-email==1.3
venusian==1.1.0
vine==1.1.4
waitress==1.1.0
WebHelpers2==2.0
WebOb==1.8.2
whichcraft==0.4.1
xlrd==1.1.0
zope.deprecation==4.3.0
zope.interface==4.5.0
zope.sqlalchemy==1.0